prototype/requirements.txt
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.8.1
Automat==0.6.0
channels==1.1.6
constantly==15.1.0
daphne==1.3.0
Django==1.11.4
django-cors-headers==2.4.0
djangorestframework==3.6.4
hyperlink==17.3.1
incremental==17.5.0
msgpack-python==0.4.8
pytz==2017.2
redis==2.10.6
six==1.10.0
Twisted==17.5.0
txaio==2.8.1
zope.interface==4.4.2
